% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.binomial_helper}
\alias{.binomial_helper}
\title{Logistic/binomial helper function}
\usage{
.binomial_helper(beta, Y, X, coi, nfolds)
}
\arguments{
\item{beta}{Fitted value.}

\item{Y}{Response vector.}

\item{X}{Data matrix.}

\item{coi}{Coefficient of interest.}

\item{nfolds}{Number of folds for cross-validation.}
}
\description{
Logistic/binomial helper function
}
\keyword{internal}
