% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.fit_lasso_wrapper}
\alias{.fit_lasso_wrapper}
\title{Lasso fit wrapper}
\usage{
.fit_lasso_wrapper(X, Y, maxit = 10000, nfolds = 5,
  family = "gaussian", lambda = NULL, refit = FALSE,
  refit_ratio = 1)
}
\description{
Lasso fit wrapper
}
