% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.gaussian_helper}
\alias{.gaussian_helper}
\title{Gaussian helper function}
\usage{
.gaussian_helper(beta, Y, X, coi, nfolds)
}
\arguments{
\item{beta}{Fitted value.}

\item{Y}{Response vector.}

\item{X}{Data matrix.}

\item{coi}{Coefficient of interest.}

\item{nfolds}{Number of folds for cross-validation.}
}
\description{
Gaussian helper function
}
\keyword{internal}
