% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.poisson_helper}
\alias{.poisson_helper}
\title{Poisson helper function}
\usage{
.poisson_helper(beta, Y, X, coi, nfolds)
}
\arguments{
\item{beta}{Fitted value.}

\item{Y}{Response vector.}

\item{X}{Data matrix.}

\item{coi}{Coefficient of interest.}

\item{nfolds}{Number of folds for cross-validation.}
}
\description{
Poisson helper function
}
\keyword{internal}
