% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dScore.R
\name{dScore}
\alias{dScore}
\title{Decorrelated Score Test.}
\usage{
dScore(Y, X, coi, beta = NULL, maxit = 1e+05, nfolds = 5,
  lambda = NULL, family = "gaussian", refit = FALSE)
}
\arguments{
\item{Y}{The response vector.}

\item{X}{The data matrix.}

\item{coi}{Coefficient of interest. It should be a single integer.}

\item{beta}{The fitted value specified by the user. The default is NULL and the function will call glmnet package for Lasso autometically.}

\item{maxit}{Maximal number of iterations. The default is 100,000.}

\item{nfolds}{Number of folds for cross validation. The default is 5.}

\item{lambda}{Tuning parameter for fitting coefficients. The default is NULL that uses default grid from glmnet.}

\item{family}{'gaussian', 'binomial' or 'poisson' for linear, logistic and poisson regression, respectively.}

\item{refit}{Logical parameter. If refit is TRUE then for coefficient estimates, the function will first use glmnet for variable selection, and then redo the regression using only the selected variables. The default is FALSE.}
}
\value{
A list containing:
\itemize{ 
\item fitted_value - The fitted coefficients using LASSO.
\item test_stats - The Score test statistic.
\item pval - P-value.
}
}
\description{
Decorrelated Score Test.
}
\examples{
X <- matrix(rnorm(10000),1000)
bt <- c(0.5,0.5,0.5,0)
Y <- X\%*\%bt + rnorm(n)
dScore(Y, X, 1,family = 'gaussian')$pval
}
